requirements.txt

sys
yfinance
pandas
sklearn.ensemble
sklearn.metrics